refactor: move styles to src
