+++ src/components/pages/parts/HeroSection/styled.tsx
@@ -1,9 +1,9 @@
 import styled from 'styled-components'
-import { typographyHeading2, typographyHeading5 } from '../../../styles/font'
+import { typographyHeading2, typographyHeading5 } from '../../../../styles/font'
 import SectionBase from '../SectionBase'
 import GalaxySvg from '../../../../../assets/bgGalaxy.svg'
-import spacing from '../../../styles/spacing'
-import mediaQuery from '../../../styles/mediaQuery'
+import spacing from '../../../../styles/spacing'
+import mediaQuery from '../../../../styles/mediaQuery'
 
 export const SectionContainer = styled(SectionBase)`
     display: grid;
