+++ src/components/pages/parts/Search/styled.ts
@@ -1,8 +1,8 @@
 import styled from 'styled-components'
-import color from '../../../styles/color'
-import { typographyParagraphBig } from '../../../styles/font'
-import spacing from '../../../styles/spacing'
-import borderRadius from '../../../styles/borderRadius'
+import color from '../../../../styles/color'
+import { typographyParagraphBig } from '../../../../styles/font'
+import spacing from '../../../../styles/spacing'
+import borderRadius from '../../../../styles/borderRadius'
 import Icon from '@mdi/react'
 
 export const SearchInput = styled.input`
