+++ src/components/pages/parts/SectionBase/styled.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
-import mediaQuery from '../../../styles/mediaQuery'
-import spacing from '../../../styles/spacing'
+import mediaQuery from '../../../../styles/mediaQuery'
+import spacing from '../../../../styles/spacing'
 
 export const SectionContainer = styled.section`
     padding: 0 ${spacing(5)};
